after merging with refactor-algorithm-page-2
